Add thread functionality for API calls and Slack listener
This commit implements two separate threads, one to make API calls to
the Mesonet API, and one to continually listen for Slack messages to
respond to. This way there is no chance that a message will get skipped
because the program is making the API call instead of listening for
messages.