Implement config.properties to protect API tokens
Previously, API tokens for the Slack integration were stored as an
environment variable, and the API token for the Mesonet API was
hardcoded into the URL string. This commit places both tokens into a
config.properties file and reads them from there, and puts
config.properties into the .gitignore, so that API keys can remain
separate and private among different users. config.sample.properties was
created as a quick guide to the structure of the config.properties. The
old API key that was hardcoded into the URL string has been disabled and
replaced with a new key in the config.properties file for security purposes.